+++ packages/insomnia/src/plugins/context/__tests__/app.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest';
 
 import appPackageJson from '../../../../package.json';
-import * as plugin from '../app-renderer';
+import * as plugin from '../app';
 
 describe('init()', () => {
